separate wheel throttling as it's own custom hook
