+++ packages/react/src/components/numberInput/NumberInput.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useRef, useState } from 'react';
 import 'hds-core';
-import throttle from 'lodash.throttle';
 import { VisuallyHidden } from '@react-aria/visually-hidden';
 
+import useThrottledWheel from '../../hooks/useThrottledWheel';
 import styles from './NumberInput.module.scss';
 import { IconMinus, IconPlus } from '../../icons';
 import { InputWrapper } from '../../internal/input-wrapper/InputWrapper';
@@ -118,29 +118,7 @@ export const NumberInput = React.forwardRef<HTMLInputElement, NumberInputProps>(
     const notifyScreenReaderStepperChangedValue = () => {
       setScreenReaderValue(String(inputRef.current.value));
     };
-    let throttledMouseWheel = false;
-
-    const throttledMouseWheeloggler = throttle(() => {
-      throttledMouseWheel = false;
-    }, 200);
-
-    useEffect(() => {
-      const ignoreScroll = (e) => {
-        if (throttledMouseWheel) {
-          e.preventDefault();
-        }
-        throttledMouseWheel = true;
-        throttledMouseWheeloggler();
-      };
-      if (inputRef.current) {
-        inputRef.current.addEventListener('wheel', ignoreScroll, { passive: false });
-      }
-      return () => {
-        if (inputRef.current) {
-          inputRef.current.removeEventListener('wheel', ignoreScroll);
-        }
-      };
-    }, [inputRef]);
+    useThrottledWheel(inputRef);
 
     /**
      * Merge props.ref to the internal ref. This is needed because we need the ref ourself and cannot rely on
